Move comment store into data store
